Disable or fix level 3 Windows compiler warnings

+++ LiteCore/Support/Error_windows.cc
@@ -78,9 +78,9 @@ namespace litecore {
 		symOptions |= SYMOPT_LOAD_LINES | SYMOPT_UNDNAME;
 		SymSetOptions(symOptions);
 		DWORD cbNeeded;
-		EnumProcessModules(process, &module_handles[0], module_handles.size() * sizeof(HMODULE), &cbNeeded);
+		EnumProcessModules(process, &module_handles[0], DWORD(module_handles.size()) * sizeof(HMODULE), &cbNeeded);
 		module_handles.resize(cbNeeded / sizeof(HMODULE));
-		EnumProcessModules(process, &module_handles[0], module_handles.size() * sizeof(HMODULE), &cbNeeded);
+		EnumProcessModules(process, &module_handles[0], DWORD(module_handles.size()) * sizeof(HMODULE), &cbNeeded);
 		transform(module_handles.begin(), module_handles.end(), back_inserter(modules), get_mod_info(process));
 
         const auto captured = CaptureStackBackTrace(0, 50, stack, nullptr);
